Merge happiness table and frontend changes

+++ happiness-backend/api/models.py
@@ -187,7 +187,7 @@ class Happiness(db.Model):
     __tablename__ = "happiness"
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     user_id = db.Column(db.Integer, db.ForeignKey("user.id"))
-    value = db.Column(db.Integer)
+    value = db.Column(db.Float)
     comment = db.Column(db.String(700))
     timestamp = db.Column(db.DateTime)
